Fix search input focus loss issue with component refactoring
- Refactor Logs page into smaller components (StatsCards, FilterPanel, LogsTable)
- Implement useTransition for non-blocking debounced updates
- Enhanced SearchInput with useLayoutEffect for robust focus handling
- Add stable component keys to prevent DOM recreation
- Improve cursor position tracking and restoration
- Memoize components to reduce unnecessary re-renders
- Add tsconfig.tsbuildinfo to .gitignore

Resolves input field losing focus when typing search queries.
Each component now has isolated re-render cycles, preventing
the search input from being recreated during state updates.
