+++ frontend/src/components/SearchInput.tsx
@@ -1,4 +1,4 @@
-import React, { memo, useRef, useEffect } from 'react'
+import React, { memo, useRef, useLayoutEffect } from 'react'
 import { Input } from '@/components/ui/input'
 import { Search } from 'lucide-react'
 
@@ -10,26 +10,38 @@ interface SearchInputProps {
 
 export const SearchInput = memo<SearchInputProps>(({ value, onChange, placeholder = "Search..." }) => {
   const inputRef = useRef<HTMLInputElement>(null)
-  const lastFocusedRef = useRef<boolean>(false)
+  const isFocusedRef = useRef<boolean>(false)
+  const lastCursorPositionRef = useRef<number>(0)
+  const lastValueRef = useRef<string>(value)
 
-  // Track if input was focused and restore it after re-renders
-  useEffect(() => {
+  // Track focus state and cursor position more reliably
+  useLayoutEffect(() => {
     const input = inputRef.current
-    if (input && lastFocusedRef.current) {
-      const cursorPosition = input.selectionStart
-      input.focus()
-      if (cursorPosition !== null) {
-        input.setSelectionRange(cursorPosition, cursorPosition)
+    if (input && isFocusedRef.current) {
+      // Only refocus if the value changed (indicating a state update)
+      // and we're not dealing with the same component instance
+      if (lastValueRef.current !== value) {
+        input.focus()
+        // Restore cursor position to the end or last known position
+        const cursorPos = Math.min(lastCursorPositionRef.current, value.length)
+        input.setSelectionRange(cursorPos, cursorPos)
       }
     }
+    lastValueRef.current = value
   })
 
-  const handleFocus = () => {
-    lastFocusedRef.current = true
+  const handleFocus = (e: React.FocusEvent<HTMLInputElement>) => {
+    isFocusedRef.current = true
+    lastCursorPositionRef.current = e.target.selectionStart || 0
   }
 
   const handleBlur = () => {
-    lastFocusedRef.current = false
+    isFocusedRef.current = false
+  }
+
+  const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    lastCursorPositionRef.current = e.target.selectionStart || 0
+    onChange(e)
   }
 
   return (
@@ -39,10 +51,12 @@ export const SearchInput = memo<SearchInputProps>(({ value, onChange, placeholde
         ref={inputRef}
         placeholder={placeholder}
         value={value}
-        onChange={onChange}
+        onChange={handleChange}
         onFocus={handleFocus}
         onBlur={handleBlur}
         className="pl-10"
+        // Add a unique key to prevent React from recreating the element
+        key="search-input-stable"
       />
     </div>
   )
